move navigation factory classes to factories package
