+++ src/main/resources/META-INF/components.txt
@@ -131,5 +131,5 @@ com.celements.configuration.CelementsAllConfigurationSource
 com.celements.pagetype.category.DefaultPageTypeCategory
 com.celements.cells.CellTypeCategory
 com.celements.web.service.LastStartupTimeStamp
-com.celements.navigation.service.XObjectNavigationFactory
-com.celements.navigation.service.JavaNavigationFactory
+com.celements.navigation.factories.XObjectNavigationFactory
+com.celements.navigation.factories.JavaNavigationFactory
