+++ src/test/java/com/celements/navigation/factories/XObjectNavigationFactoryTest.java
@@ -1,4 +1,4 @@
-package com.celements.navigation.service;
+package com.celements.navigation.factories;
 
 import static com.celements.common.test.CelementsTestUtils.*;
 import static org.easymock.EasyMock.*;
@@ -15,6 +15,8 @@ import com.celements.common.test.AbstractComponentTest;
 import com.celements.navigation.INavigation;
 import com.celements.navigation.INavigationClassConfig;
 import com.celements.navigation.NavigationConfig;
+import com.celements.navigation.factories.NavigationFactory;
+import com.celements.navigation.factories.XObjectNavigationFactory;
 import com.celements.navigation.presentation.IPresentationTypeRole;
 import com.celements.web.service.IWebUtilsService;
 import com.xpn.xwiki.XWiki;
